pyproject.toml
[project]
name = "ratatouille"
version = "2.0.0"
description = "🐀 Self-hostable data platform - Anyone Can Data!"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Ratatouille Team"}
]
keywords = ["data", "platform", "lakehouse", "duckdb", "iceberg"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
]

dependencies = [
    "duckdb>=1.0.0",
    "pyiceberg[s3]>=0.7.0",
    "pynessie>=0.67.0",
    "pandas>=2.0.0",
    "pyarrow>=15.0.0",
    "polars>=0.20.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "packaging>=24.0",
    "httpx>=0.27.0",
    "boto3>=1.34.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.31.0",
    "minio>=7.2.0",
]
notebooks = [
    "jupyterlab>=4.0.0",
    "ipykernel>=6.0.0",
]
dagster = [
    "dagster>=1.6.0",
    "dagster-webserver>=1.6.0",
]

[project.scripts]
rat = "ratatouille.cli:main"

[project.urls]
Homepage = "https://github.com/ratatouille-data/ratatouille"
Documentation = "https://github.com/ratatouille-data/ratatouille/docs"
Repository = "https://github.com/ratatouille-data/ratatouille"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.dagster]
module_name = "ratatouille.definitions"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
]
ignore = ["E501"]  # line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests (require services)",
    "slow: marks tests as slow",
]

[tool.coverage.run]
source = ["src/ratatouille"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true
